+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 August 2025
 
+* Make it possible to respawn on death
+* Remove some print-to-file functionality on death (tombstone and character sheet). If I miss them in the future I'll add a command for them instead.
+
 ## 5-alpha524
 
 August 2025